requirements.txt
altair==4.1.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1613037829118/work
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///opt/conda/conda-bld/attrs_1642510447205/work
backcall @ file:///C:/miniconda3/conda-bld/backcall_1761744905541/work
backports.zoneinfo==0.2.1
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
blinker==1.5
cachetools==4.2.4
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1625831763871/work
charset-normalizer==2.0.12
click==8.0.4
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
comm==0.1.4
commonmark==0.9.1
cycler==0.11.0
dataclasses==0.8
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dlib @ https://files.pythonhosted.org/packages/0e/ce/f8a3cff33ac03a8219768f0694c5d703c8e037e6aba2e865f9bae22ed63c/dlib-19.8.1-cp36-cp36m-win_amd64.whl#sha256=794994fa2c54e7776659fddb148363a5556468a6d5d46be8dad311722d54bfcf
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
gitdb==4.0.9
GitPython==3.1.18
idna==3.10
importlib-metadata @ file:///C:/ci/importlib-metadata_1631916885413/work
importlib-resources==5.4.0
ipykernel @ file:///C:/ci/ipykernel_1596206775157/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593446240034/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets==7.8.5
jedi @ file:///C:/ci/jedi_1606914470086/work
Jinja2 @ file:///opt/conda/conda-bld/jinja2_1647436528585/work
jsonschema @ file:///Users/ktietz/demo/mc3/conda-bld/jsonschema_1630511932244/work
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-core @ file:///C:/ci/jupyter_core_1633402348205/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab_widgets==1.1.11
kiwisolver==1.3.1
MarkupSafe @ file:///C:/ci/markupsafe_1621528313524/work
matplotlib==3.3.4
mistune==0.8.4
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///C:/ci/nbconvert_1601896775581/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
notebook @ file:///C:/ci/notebook_1629205950149/work
numpy @ file:///D:/bld/numpy_1626682185159/work
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.1.5
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso==0.7.0
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
prometheus-client @ file:///opt/conda/conda-bld/prometheus_client_1643788673601/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.6
pyarrow==6.0.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.6.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
Pympler==1.1
pyparsing @ file:///tmp/build/80754af9/pyparsing_1635766073266/work
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pywin32==228
pywinpty==0.5.7
pyzmq @ file:///C:/ci/pyzmq_1628276569449/work
requests==2.27.1
rich==12.6.0
semver==2.13.0
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
streamlit==1.10.0
terminado==0.9.4
testpath @ file:///tmp/build/80754af9/testpath_1624638946665/work
toml==0.10.2
toolz==0.12.0
tornado @ file:///C:/ci/tornado_1606942379977/work
traitlets @ file:///C:/ci/traitlets_1632759765830/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
tzdata==2025.2
tzlocal==4.2
urllib3==1.26.20
validators==0.20.0
watchdog==2.3.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
widgetsnbextension==3.6.10
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1633618647012/work
